[TIP] Rename lib folders to utils
