+++ x-pack/plugins/threat_intelligence/public/modules/indicators/components/indicators_flyout/tabs/indicators_flyout_overview/components/highlighted_values_table/highlighted_values_table.tsx
@@ -7,7 +7,7 @@
 
 import React, { useMemo, VFC } from 'react';
 import { Indicator, RawIndicatorFieldId } from '../../../../../../../../../common/types/indicator';
-import { unwrapValue } from '../../../../../../lib/unwrap_value';
+import { unwrapValue } from '../../../../../../utils/unwrap_value';
 import { IndicatorFieldsTable } from '../../../../components/indicator_fields_table';
 
 /**
