+++ x-pack/plugins/threat_intelligence/public/modules/indicators/hooks/use_indicators.ts
@@ -18,8 +18,8 @@ import { useInspector } from '../../../hooks/use_inspector';
 import { Indicator } from '../../../../common/types/indicator';
 import { useKibana } from '../../../hooks/use_kibana';
 import { useSourcererDataView } from './use_sourcerer_data_view';
-import { getRuntimeMappings } from '../lib/get_runtime_mappings';
-import { getIndicatorsQuery } from '../lib/get_indicators_query';
+import { getRuntimeMappings } from '../utils/get_runtime_mappings';
+import { getIndicatorsQuery } from '../utils/get_indicators_query';
 
 const PAGE_SIZES = [10, 25, 50];
