+++ x-pack/plugins/threat_intelligence/public/modules/indicators/utils/display_name.ts
