+++ x-pack/plugins/threat_intelligence/public/modules/indicators/utils/field_value.ts
