+++ x-pack/plugins/threat_intelligence/public/modules/indicators/utils/get_runtime_mappings.ts
