+++ x-pack/plugins/threat_intelligence/public/modules/query_bar/utils/filter.ts
