+++ x-pack/plugins/threat_intelligence/public/modules/timeline/components/add_to_timeline/add_to_timeline.tsx
@@ -12,9 +12,12 @@ import { EuiButtonEmpty, EuiButtonIcon } from '@elastic/eui/src/components/butto
 import { EuiContextMenuItem, EuiFlexItem, EuiToolTip } from '@elastic/eui';
 import { i18n } from '@kbn/i18n';
 import { FormattedMessage } from '@kbn/i18n-react';
-import { generateDataProvider } from '../../lib/data_provider';
+import { generateDataProvider } from '../../utils/data_provider';
 import { ComponentType } from '../../../../../common/types/component_type';
-import { fieldAndValueValid, getIndicatorFieldAndValue } from '../../../indicators/lib/field_value';
+import {
+  fieldAndValueValid,
+  getIndicatorFieldAndValue,
+} from '../../../indicators/utils/field_value';
 import { useKibana } from '../../../../hooks/use_kibana';
 import { Indicator } from '../../../../../common/types/indicator';
 import { useStyles } from './styles';
